Work on: CVS BS analysis I-file Reading Various other stuff...
Former-commit-id: 360d175010396ee4a91c69383e69a991612743e1 [formerly 808f0268c008d68a474b3af89ec0f57adf3542a3]
Former-commit-id: efa46f52b431e674eee4aed4a2b5f41e02519c45
Former-commit-id: b93b1600e142c4eb212cf5d990e05e1ec0771463
Former-commit-id: c2a37693832a267dc94b3194339f4540d0908235 [formerly bd3cd2ba5bd154ba434a2712279b8bd74c4d540b]
Former-commit-id: cb9b4e3d22d2e5a55acfd87e1c2ef922e6aa3159
Former-commit-id: f0a908ab906757e9e73860998e47f96e01c66522

+++ @ac_data_cl/get_datamat.m
@@ -4,7 +4,7 @@ function [datamat,idx]=get_datamat(data,field)
 [idx,found]=find_field_idx(data,(deblank(field)));
 
 if found
-    datamat=single(data.SubData(idx).Memap.Data.(lower(deblank(field))));
+    datamat=double(data.SubData(idx).Memap.Data.(lower(deblank(field))));
 else
     datamat=[];
 end
